server/requirements.txt
Flask
Flask-Cors
protobuf==3.20.0
requests
tensorflow==2.8.0
pandas
google_api_python_client